# paas-mediation-service/docs/swagger.yaml
definitions:
...
  v2.ContainerEnv:
    properties:
      name:
        type: string
      value:
        type: string
      valueFrom:
        $ref: '#/definitions/v2.ValueFrom'
    type: object
  v2.ContainerPort:
    properties:
      containerPort:
        type: integer
      name:
        type: string
      protocol:
        type: string
    type: object
  v2.ContainerResources:
    properties:
      limits:
        $ref: '#/definitions/v2.CpuMemoryResource'
      requests:
        $ref: '#/definitions/v2.CpuMemoryResource'
    type: object
  v2.ContainerState:
    properties:
      running:
        $ref: '#/definitions/v2.ContainerStateRunning'
      terminated:
        $ref: '#/definitions/v2.ContainerStateTerminated'
      waiting:
        $ref: '#/definitions/v2.ContainerStateWaiting'
    type: object
  v2.ContainerStateRunning:
    properties:
      startedAt:
        type: string
    type: object
  v2.ContainerStateTerminated:
    properties:
      containerID:
        type: string
      exitCode:
        type: integer
      finishedAt:
        type: string
      reason:
        type: string
      startedAt:
        type: string
    type: object
  v2.ContainerStateWaiting:
    properties:
      message:
        type: string
      reason:
        type: string
    type: object
  v2.ContainerStatus:
    properties:
      containerID:
        type: string
      image:
        type: string
      imageID:
        type: string
      lastState:
        $ref: '#/definitions/v2.ContainerState'
      name:
        type: string
      ready:
        type: boolean
      restartCount:
        type: integer
      state:
        $ref: '#/definitions/v2.ContainerState'
    type: object
  v2.ContainerVolumeMount:
    properties:
      mountPath:
        type: string
      name:
        type: string
      readOnly:
        type: boolean
    type: object
  v2.CpuMemoryResource:
    properties:
      cpu:
        type: string
      memory:
        type: string
    type: object
  v2.Deployment:
    properties:
      metadata:
        $ref: '#/definitions/v2.Metadata'
      spec:
        $ref: '#/definitions/v2.DeploymentSpec'
      status:
        $ref: '#/definitions/v2.DeploymentStatus'
    type: object
  v2.DeploymentCondition:
    properties:
      lastTransitionTime:
        type: string
      lastUpdateTime:
        type: string
      message:
        type: string
      reason:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
...
  v2.DeploymentSpec:
    properties:
      replicas:
        type: integer
      revisionHistoryLimit:
        type: integer
      strategy:
        $ref: '#/definitions/v2.DeploymentStrategy'
      template:
        $ref: '#/definitions/v2.PodTemplateSpec'
    type: object
  v2.DeploymentStatus:
    properties:
      availableReplicas:
        type: integer
      conditions:
        items:
          $ref: '#/definitions/v2.DeploymentCondition'
        type: array
      observedGeneration:
        type: integer
      readyReplicas:
        type: integer
      replicas:
        type: integer
      updatedReplicas:
        type: integer
    type: object
  v2.DeploymentStrategy:
    properties:
      type:
        type: string
    type: object
  v2.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  v2.FieldRef:
    properties:
      apiVersion:
        type: string
      fieldPath:
        type: string
    type: object
  v2.Metadata:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      generation:
        type: integer
      kind:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
      namespace:
        type: string
      resourceVersion:
        type: string
      uid:
        type: string
    type: object
...
  v2.Pod:
    properties:
      metadata:
        $ref: '#/definitions/v2.Metadata'
      spec:
        $ref: '#/definitions/v2.PodSpec'
      status:
        $ref: '#/definitions/v2.PodStatus'
    type: object
  v2.PodSpec:
    properties:
      containers:
        items:
          $ref: '#/definitions/v2.SpecContainer'
        type: array
      dnsPolicy:
        type: string
      nodeName:
        type: string
      restartPolicy:
        type: string
      terminationGracePeriodSeconds:
        type: integer
      volumes:
        items:
          $ref: '#/definitions/v2.SpecVolume'
        type: array
    type: object
  v2.PodStatus:
    properties:
      conditions:
        items:
          $ref: '#/definitions/v2.StatusCondition'
        type: array
      containerStatuses:
        items:
          $ref: '#/definitions/v2.ContainerStatus'
        type: array
      hostIP:
        type: string
      phase:
        type: string
      podIP:
        type: string
      startTime:
        type: string
    type: object
  v2.PodTemplateSpec:
    properties:
      metadata:
        $ref: '#/definitions/v2.TemplateMetadata'
      spec:
        $ref: '#/definitions/v2.PodSpec'
    type: object
...
  v2.Route:
    properties:
      metadata:
        $ref: '#/definitions/v2.Metadata'
      spec:
        $ref: '#/definitions/v2.RouteSpec'
    type: object
  v2.RoutePort:
    properties:
      targetPort:
        type: integer
    type: object
  v2.RouteSpec:
    properties:
      host:
        type: string
      ingressClassName:
        type: string
      path:
        type: string
      pathType:
        type: string
      port:
        $ref: '#/definitions/v2.RoutePort'
      to:
        $ref: '#/definitions/v2.Target'
    type: object
  v2.SecretKeyRef:
    properties:
      key:
        type: string
      name:
        type: string
    type: object
...
  v2.SpecContainer:
    properties:
      args:
        items:
          type: string
        type: array
      env:
        items:
          $ref: '#/definitions/v2.ContainerEnv'
        type: array
      image:
        type: string
      imagePullPolicy:
        type: string
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/v2.ContainerPort'
        type: array
      resources:
        $ref: '#/definitions/v2.ContainerResources'
      volumeMounts:
        items:
          $ref: '#/definitions/v2.ContainerVolumeMount'
        type: array
    type: object
  v2.SpecVolume:
    properties:
      name:
        type: string
      secret:
        $ref: '#/definitions/v2.VolumesSecret'
    type: object
  v2.StatusCondition:
    properties:
      lastProbeTime:
        type: string
      lastTransitionTime:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  v2.Target:
    properties:
      name:
        type: string
    type: object
  v2.TemplateMetadata:
    properties:
      labels:
        additionalProperties:
          type: string
        type: object
    type: object
  v2.ValueFrom:
    properties:
      fieldRef:
        $ref: '#/definitions/v2.FieldRef'
      secretKeyRef:
        $ref: '#/definitions/v2.SecretKeyRef'
    type: object
  v2.VolumesSecret:
    properties:
      defaultMode:
        type: integer
      secretName:
        type: string
    type: object
info:
  description: API for Paas Mediation.
  title: Paas Mediation API
  version: "2.0"
paths:
...
  /api/v2/namespaces/{namespace}/deployments:
    get:
      consumes:
      - application/json
      description: Get Deployment by name and namespace
      operationId: v2-get-deployment-list
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: query
        name: annotations
        type: string
      - description: resource name
        in: query
        name: labels
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v2.Deployment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Deployment by name and namespace
      tags:
      - since:2.0
  /api/v2/namespaces/{namespace}/deployments/{name}:
    get:
      consumes:
      - application/json
      description: Get Deployment by name and namespace
      operationId: v2-get-deployment
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Deployment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Deployment by name and namespace
      tags:
      - since:2.0
  /api/v2/namespaces/{namespace}/pods:
    get:
      consumes:
      - application/json
      description: Get Pod by name and namespace
      operationId: v2-get-pod-list
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: query
        name: annotations
        type: string
      - description: resource name
        in: query
        name: labels
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v2.Pod'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Pod by name and namespace
      tags:
      - since:2.0
  /api/v2/namespaces/{namespace}/pods/{name}:
    get:
      consumes:
      - application/json
      description: Get Pod by name and namespace
      operationId: v2-get-pod
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Pod'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Pod by name and namespace
      tags:
      - since:2.0
...
  /api/v2/namespaces/{namespace}/routes:
    get:
      consumes:
      - application/json
      description: Get Route by name and namespace
      operationId: v2-get-route-list
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: query
        name: annotations
        type: string
      - description: resource name
        in: query
        name: labels
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/v2.Route'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Route by name and namespace
      tags:
      - since:2.0
    post:
      consumes:
      - application/json
      description: Create Route in namespace
      operationId: v2-create-route
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v2.Route'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Route'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create Route in namespace
      tags:
      - since:2.0
    put:
      consumes:
      - application/json
      description: Update or Create Route in namespace
      operationId: v2-update-or-create-route
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v2.Route'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Route'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update or Create Route in namespace
      tags:
      - since:2.0
  /api/v2/namespaces/{namespace}/routes/{name}:
    delete:
      consumes:
      - application/json
      description: Delete Route with name in namespace
      operationId: v2-delete-route
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Route'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete Route with name in namespace
      tags:
      - since:2.0
    get:
      consumes:
      - application/json
      description: Get Route by name and namespace
      operationId: v2-get-route
      parameters:
      - description: target namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: resource name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v2.Route'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v2.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get Route by name and namespace
      tags:
      - since:2.0
...
produces:
- application/json
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...